requirements.txt
amqp==2.5.2
asgi-redis==1.4.3
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
backcall==0.1.0
billiard==3.6.3.0
bleach==3.1.1
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
coverage==5.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.15
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.7
django-background-tasks==1.2.5
django-chartit==0.2.9
django-compat==1.0.15
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-extensions==2.2.8
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-framework==0.1.0
django-simple-history==2.8.0
django-summernote==0.8.11.6
django-taggit==1.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
flower==0.9.4
graphviz==0.14
humanize==0.5.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
kombu==4.6.8
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
msgpack-python==0.5.6
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.0
packaging==19.2
pandocfilters==1.4.2
parso==0.6.2
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pyflakes==2.1.1
Pygments==2.6.1
PyHamcrest==2.0.2
pymemcache==3.1.1
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.7
pytest==5.3.2
pytest-django==3.7.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
redis==2.10.6
reportlab==3.5.32
requests==2.23.0
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
stevedore==1.32.0
supervisor==4.2.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.8
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zope.interface==5.1.0
